+++ CHANGELOG.md
@@ -92,6 +92,7 @@
       - LibNotify -- Uses the libnotify library to send notifications for each discovered issue and a summary at the end of the scan.
       - EmailNotify -- Sends a notification (and optionally a report) over SMTP at the end of the scan.
       - Manual verification -- Flags issues that require manual verification as untrusted in order to reduce the signal-to-noise ratio.
+      - Resolver -- Resolves vulnerable hostnames to IP addresses.
 - Reports
    - HTML report
       - Fixed replay forms to include URL params in the <em>action</em> attribute. [Issue #73]